finish backebd product + dev front product + dev cart session

+++ src/Controller/AdminController.php
@@ -32,6 +32,16 @@ final class AdminController extends AbstractController
     {
         // ?Product $product : le ? veut dire que par défault $product a une valeur null
         // dump($product);
+
+        // 1er route '/admin/products'
+
+        // Si la variable $product N'EST PAS (!), si elle renvoie false, cela veut dire qu'aucun id product n'est passé dans l'URL, alors on entre dans la condition, et on initialise un objet Entity $product, donc c'est une insertion de produit 
+
+
+
+        // 2ème route : '/admin/products/update/{id}'
+
+        // On envoi un id $product dans l'URL, Symfony comprend que l'on a besoin d'un objet entity product issu de la table SQL product, il est capable automatiquement d'aller sélectionner en BDD le produit et de l'envoyer en arguement de la fonction ?Product $product, à ce moment là, la variable $product contient les données du produit que l'on souhaite modifié, alors on ne rentre pas dans la condition pas dans la condition if
         if (!$product)
             $product = new Product;
 
@@ -68,6 +78,14 @@ final class AdminController extends AbstractController
                 // dump($product);
             }
 
+            // Si la condition retourne TRUE, cela veut dire que l'id est connu en BDD, c'est une modification.
+            if ($product->getId()) {
+                $messageValidate = "Les modification ont été enregistrées.";
+            } else {
+                // Sinon dans tous les autres cas, c'est une insertion
+                $messageValidate = "L'article à été enregistrées.";
+            }
+
             $product->setCreatedAt(new \DateTimeImmutable());
             $entityManager->persist($product);
             $entityManager->flush();
@@ -86,6 +104,23 @@ final class AdminController extends AbstractController
             'dbProducts' => $dbProducts,
             'pictureFile' => $product->getPicture()
         ]);
+    }
+
+    #[Route('/admin/products/remove/{id}', name: 'app_admin_products_remove')]
+    public function adminRemoveProduct($id, ProductRepository $repoProduct, EntityManagerInterface $entityManager)
+    {
+        // SELECT * FROM product WHERE id = $id
+        $product = $repoProduct->find($id);
+        // dump($product);
+
+        // DELETE FROM product WHERE id = x
+        $entityManager->remove($product);
+        // execute();
+        $entityManager->flush();
+
+        $this->addFlash('success', "L'article a été supprimé.");
+
+        return $this->redirectToRoute('app_admin_products');
     }
 
     #[Route('/admin/category', name: 'app_admin_category')]
@@ -163,13 +198,17 @@ final class AdminController extends AbstractController
     public function adminCategoryRemove($id, EntityManagerInterface $entityManager, CategoryRepository $repoCategory)
     {
         $category = $repoCategory->find($id);
-        dump($category);
+        dump($category->getProducts()->isEmpty());
 
-        // DELETE FROM category WHERE id = $id;
-        $entityManager->remove($category);
-        $entityManager->flush();
+        if ($category->getProducts()->isEmpty()) {
+            // DELETE FROM category WHERE id = $id;
+            $entityManager->remove($category);
+            $entityManager->flush();
 
-        $this->addFlash('success', "La catégorie a bien été supprimée.");
+            $this->addFlash('success', "La catégorie a bien été supprimée.");
+        } else {
+            $this->addFlash('danger', "Impossible de supprimer la catégorie, les articles y sont associés.");
+        }
 
         return $this->redirectToRoute('app_admin_category');
     }
